+++ src/data.hpp
@@ -1,0 +1,12 @@
+struct ComparisonConfig {
+    bool isBuffed;
+    float sawSpeed;
+
+    //bool mapGroups;
+    //bool markDBlocks;
+    //bool mapTriggers;
+};
+
+std::vector<int> groupKeys = {
+    33, 57, 274, 108, 51, 71, 76, 395, 442, 516, 517, 518, 519, 448, 455, 457
+};